test: :white_check_mark: Debug unit tests; add example test for CardActivity
