+++ src/components/CardActivity.js
@@ -1,18 +1,11 @@
 import React, { useMemo } from "react";
 import styled from "@emotion/styled";
 import { useQuery } from "react-query";
-import {
-  useTable,
-  useSortBy,
-  useAsyncDebounce,
-  useFilters,
-  useGlobalFilter,
-} from "react-table";
-import { matchSorter } from "match-sorter";
 import moment from "moment";
 
 import { getCardActivity } from "api/card";
 import { H2 } from "components/lib/typography";
+import Table from "components/lib/Table";
 import { getUserCardId } from "components/CardOverview";
 
 function CardActivity() {
@@ -21,7 +14,7 @@ function CardActivity() {
     data: transactions,
     isLoading,
     isError,
-  } = useQuery("CardActivity", () => getCardActivity(cardId));
+  } = useQuery("CardActivity", () => getCardActivity(cardId), { retry: false });
   const data = useMemo(
     () =>
       transactions &&
@@ -47,6 +40,13 @@ function CardActivity() {
   return (
     <Container>
       <H2>Activity</H2>
+      {isLoading && <span>Loading...</span>}
+      {isError && (
+        <span>
+          A server error occurred. Try refreshing or contact management for
+          assistance.
+        </span>
+      )}
       {!isLoading && !isError && <Table columns={columns} data={data} />}
     </Container>
   );
@@ -62,187 +62,3 @@ const Container = styled.div`
   box-sizing: border-box;
 `;
 export default CardActivity;
-
-// Table stuff starts here
-// Let the table remove the filter if the string is empty
-fuzzyTextFilterFn.autoRemove = (val) => !val;
-function Table({ columns, data }) {
-  const filterTypes = React.useMemo(
-    () => ({
-      fuzzyText: fuzzyTextFilterFn,
-    }),
-    []
-  );
-
-  const defaultColumn = useMemo(
-    () => ({
-      Filter: DefaultColumnFilter,
-    }),
-    []
-  );
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    state,
-    rows,
-    prepareRow,
-    preGlobalFilteredRows,
-    setGlobalFilter,
-  } = useTable(
-    {
-      columns,
-      data,
-      defaultColumn,
-      filterTypes,
-    },
-    useFilters,
-    useGlobalFilter,
-    useSortBy
-  );
-  return (
-    <TableContainer>
-      <GlobalFilter
-        preGlobalFilteredRows={preGlobalFilteredRows}
-        globalFilter={state.globalFilter}
-        setGlobalFilter={setGlobalFilter}
-      />
-      <table
-        {...getTableProps()}
-        style={{
-          width: "100%",
-          textAlign: "left",
-        }}
-      >
-        <thead style={{ height: "80px", boxSizing: "content-box" }}>
-          {headerGroups.map((headerGroup) => (
-            <tr {...headerGroup.getHeaderGroupProps()}>
-              {headerGroup.headers.map((column) => (
-                <th
-                  {...column.getHeaderProps(column.getSortByToggleProps())}
-                  style={{
-                    minWidth: "120px",
-                    lineHeight: "20px",
-                  }}
-                >
-                  {column.render("Header")}
-                  <span>
-                    {column.isSorted
-                      ? column.isSortedDesc
-                        ? " 🔽"
-                        : " 🔼"
-                      : ""}
-                  </span>
-                  <div>{column.canFilter ? column.render("Filter") : null}</div>
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-        <tbody {...getTableBodyProps()}>
-          {rows.map((row, i) => {
-            prepareRow(row);
-            return (
-              <tr {...row.getRowProps()}>
-                {row.cells.map((cell) => {
-                  return (
-                    <td {...cell.getCellProps()}>{cell.render("Cell")}</td>
-                  );
-                })}
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
-    </TableContainer>
-  );
-}
-const TableContainer = styled.div`
-  height: 300px;
-  width: 1000px;
-  margin: 0 auto;
-  overflow: scroll;
-  padding: 24px;
-
-  table {
-    border-spacing: 0;
-    border: 1px solid black;
-    tr {
-      :last-child {
-        td {
-          border-bottom: 0;
-        }
-      }
-    }
-
-    th,
-    td {
-      margin: 0;
-      padding: 0.5rem;
-      border-bottom: 1px solid black;
-      border-right: 1px solid black;
-
-      :last-child {
-        border-right: 0;
-      }
-    }
-  }
-`;
-
-function GlobalFilter({
-  preGlobalFilteredRows,
-  globalFilter,
-  setGlobalFilter,
-}) {
-  const count = preGlobalFilteredRows.length;
-  const [value, setValue] = React.useState(globalFilter);
-  const onChange = useAsyncDebounce((value) => {
-    setGlobalFilter(value || undefined);
-  }, 200);
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        alignItems: "flex-end",
-        marginBottom: "18px",
-      }}
-    >
-      Global Search:
-      <input
-        value={value || ""}
-        onChange={(e) => {
-          setValue(e.target.value);
-          onChange(e.target.value);
-        }}
-        placeholder={`${count} records...`}
-        style={{
-          fontSize: "1.1rem",
-          border: "0",
-          marginTop: "8px",
-          marginLeft: "8px",
-        }}
-      />
-    </div>
-  );
-}
-
-function DefaultColumnFilter({
-  column: { filterValue, preFilteredRows, setFilter },
-}) {
-  const count = preFilteredRows.length;
-
-  return (
-    <input
-      value={filterValue || ""}
-      onChange={(e) => {
-        setFilter(e.target.value || undefined); // Set undefined to remove the filter entirely
-      }}
-      placeholder={`Search ${count} records...`}
-    />
-  );
-}
-
-function fuzzyTextFilterFn(rows, id, filterValue) {
-  return matchSorter(rows, filterValue, { keys: [(row) => row.values[id]] });
-}
