+++ src/setupTests.js
@@ -2,6 +2,7 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
+import "react-app-polyfill/stable";
 import "@testing-library/jest-dom";
 import { server } from "./mocks/server.js";
